requirements.txt
torch>=1.8.0
torchvision>=0.9.0
ultralytics>=8.0.0
paddlepaddle>=2.4.0
paddleocr==2.7.3
lmdb==1.4.1
opencv-python>=4.5.0
numpy>=1.20.0
Pillow>=8.0.0
streamlit>=1.20.0
tqdm
matplotlib
pandas
